본인 게시글만 수정 삭제 가능 기능 추가

+++ carrot/src/main/java/com/market/carrot/Controller/ApiController.java
@@ -1,5 +1,8 @@
 package com.market.carrot.Controller;
 
+import com.market.carrot.domain.Item;
+import com.market.carrot.domain.user.User;
+import com.market.carrot.dto.ItemDto;
 import com.market.carrot.dto.ItemSaveDto;
 import com.market.carrot.dto.ItemUpdateDto;
 import com.market.carrot.dto.UserSaveDto;
@@ -10,6 +13,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
+import java.util.Optional;
+
 @RestController
 @RequiredArgsConstructor
 public class ApiController {
@@ -18,19 +24,37 @@ public class ApiController {
     private final UserService userService;
 
     @PostMapping("/api/v1/items")
-    public Long save(@RequestBody ItemSaveDto itemSaveDto){
+    public Long save(@RequestBody ItemSaveDto itemSaveDto, Principal principal){
+        User user = this.userService.getUser(principal.getName());
+        itemSaveDto.setOwner(user);
         return itemsService.save(itemSaveDto);
     }
 
     @PutMapping("/api/v1/items/{id}")
-    private Long update(@PathVariable Long id, @RequestBody ItemUpdateDto updateDto){
-        return itemsService.update(id,updateDto);
+    private Long update(@PathVariable Long id, @RequestBody ItemUpdateDto updateDto,Principal principal){
+        ItemDto itemOwner = this.itemsService.findById(id);
+        User user = this.userService.getUser(principal.getName());
+        if(itemOwner.getOwner().getId()==user.getId()){
+            return itemsService.update(id,updateDto);
+        }
+        else{
+            throw new IllegalStateException("본인의 게시글이 아닌 게시글을 수정할 수 없습니다.");
+        }
+
     }
 
     @DeleteMapping("/api/v1/items/{id}")
-    private Long delete(@PathVariable Long id){
-        itemsService.delete(id);
-        return id;
+    private Long delete(@PathVariable Long id, Principal principal){
+        ItemDto itemOwner = this.itemsService.findById(id);
+        User user = this.userService.getUser(principal.getName());
+        if(itemOwner.getOwner().getId()==user.getId()){
+            itemsService.delete(id);
+            return id;
+        }
+        else{
+            throw new IllegalStateException("본인의 게시글이 아닌 게시글은 삭제할 수 없습니다.");
+        }
+
     }
 
     @PostMapping("/api/v1/join")
